fix problem with time zone information in libgerrit
If tz was lower than 0 the complete ChangeInfo was not recognized

+++ libgerrit/entities.rs
@@ -169,7 +169,7 @@ pub struct GitPersonInfo {
     /// The timestamp of when this identity was constructed.
     pub date: String,
     /// The timezone offset from UTC of when this identity was constructed.
-    pub tz: u16,
+    pub tz: i32,
 }
 
 /// `CommitInfoParents`, same as `CommitInfo` but commit is string
